+++ include/timer.h
@@ -7,4 +7,4 @@
 #define TYPE_SUNDOWN 2
 
 void timerCyclic();
-void getSunriseOrSunset(uint8_t type, int8_t offset, float latitude, float longitude, uint8_t &hour, uint8_t &minute);
+void getSunriseOrSunset(uint8_t type, int16_t offset, float latitude, float longitude, uint8_t &hour, uint8_t &minute);
